feat: Add TracingConfig for configuration settings in instrumentations
